runWithMocha in a file
